feat: Exception 클래스 API 멀티모듈로 이동
