+++ Api/src/main/java/com/mapz/api/common/exception/common/FailConvertToJsonException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.common;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class FailConvertToJsonException extends CustomException {
+    public FailConvertToJsonException() {
+        super(ExceptionDetails.FAIL_CONVERT_TO_JSON);
+    }
+}
