+++ Api/src/main/java/com/mapz/api/common/exception/common/FailParseException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.common;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class FailParseException extends CustomException {
+    public FailParseException(Throwable cause) {
+        super(ExceptionDetails.FAIL_PARSE, cause);
+    }
+}
