+++ Api/src/main/java/com/mapz/api/common/exception/diary/AlreadyLikedDiaryException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.diary;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class AlreadyLikedDiaryException extends CustomException {
+    public AlreadyLikedDiaryException() {
+        super(ExceptionDetails.ALREADY_LIKED_DIARY);
+    }
+}
