+++ Api/src/main/java/com/mapz/api/common/exception/jwt/JwtExpiredException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.jwt;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class JwtExpiredException extends CustomException {
+    public JwtExpiredException(Throwable cause) {
+        super(ExceptionDetails.EXPIRED_TOKEN, cause);
+    }
+}
