+++ Api/src/main/java/com/mapz/api/common/exception/report/AlreadyReportedDiary.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.report;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class AlreadyReportedDiary extends CustomException {
+    public AlreadyReportedDiary() {
+        super(ExceptionDetails.ALREADY_REPORTED_DIARY);
+    }
+}
