+++ Api/src/main/java/com/mapz/api/common/exception/user/DuplicatedEmailException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.user;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class DuplicatedEmailException extends CustomException {
+    public DuplicatedEmailException() {
+        super(ExceptionDetails.DUPLICATED_EMAIL);
+    }
+}
