+++ Api/src/main/java/com/mapz/api/common/exception/user/NoPermissionUserException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.user;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class NoPermissionUserException extends CustomException {
+    public NoPermissionUserException() {
+        super(ExceptionDetails.NO_PERMISSION_USER);
+    }
+}
