+++ Api/src/main/java/com/mapz/api/common/exception/user/PresentUserException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.user;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class PresentUserException extends CustomException {
+    public PresentUserException() {
+        super(ExceptionDetails.PRESENT_USER);
+    }
+}
