+++ Api/src/main/java/com/mapz/api/common/exception/user/WrongPasswordException.java
@@ -1,0 +1,10 @@
+package com.mapz.api.common.exception.user;
+
+import com.mapz.api.common.exception.CustomException;
+import com.mapz.api.common.exception.ExceptionDetails;
+
+public class WrongPasswordException extends CustomException {
+    public WrongPasswordException() {
+        super(ExceptionDetails.WRONG_PASSWORD);
+    }
+}
