--- src/main/java/com/cheocharm/MapZ/common/exception/S3/FailConvertToFileException.java
@@ -1,10 +1,0 @@
-package com.cheocharm.MapZ.common.exception.S3;
-
-import com.cheocharm.MapZ.common.exception.CustomException;
-import com.cheocharm.MapZ.common.exception.ExceptionDetails;
-
-public class FailConvertToFileException extends CustomException {
-    public FailConvertToFileException() {
-        super(ExceptionDetails.FAIL_CONVERT_TO_FILE);
-    }
-}
