--- src/main/java/com/cheocharm/MapZ/common/exception/common/FailConvertToJsonException.java
@@ -1,10 +1,0 @@
-package com.cheocharm.MapZ.common.exception.common;
-
-import com.cheocharm.MapZ.common.exception.CustomException;
-import com.cheocharm.MapZ.common.exception.ExceptionDetails;
-
-public class FailConvertToJsonException extends CustomException {
-    public FailConvertToJsonException() {
-        super(ExceptionDetails.FAIL_CONVERT_TO_JSON);
-    }
-}
