--- src/main/java/com/cheocharm/MapZ/common/exception/user/DuplicatedUsernameException.java
@@ -1,10 +1,0 @@
-package com.cheocharm.MapZ.common.exception.user;
-
-import com.cheocharm.MapZ.common.exception.CustomException;
-import com.cheocharm.MapZ.common.exception.ExceptionDetails;
-
-public class DuplicatedUsernameException extends CustomException {
-    public DuplicatedUsernameException() {
-        super(ExceptionDetails.DUPLICATED_USERNAME);
-    }
-}
