--- src/main/java/com/cheocharm/MapZ/common/exception/user/WrongPasswordException.java
@@ -1,10 +1,0 @@
-package com.cheocharm.MapZ.common.exception.user;
-
-import com.cheocharm.MapZ.common.exception.CustomException;
-import com.cheocharm.MapZ.common.exception.ExceptionDetails;
-
-public class WrongPasswordException extends CustomException {
-    public WrongPasswordException() {
-        super(ExceptionDetails.WRONG_PASSWORD);
-    }
-}
