--- src/main/java/com/cheocharm/MapZ/common/exception/usergroup/NotFoundUserGroupException.java
@@ -1,11 +1,0 @@
-package com.cheocharm.MapZ.common.exception.usergroup;
-
-import com.cheocharm.MapZ.common.exception.CustomException;
-import com.cheocharm.MapZ.common.exception.ExceptionDetails;
-
-public class NotFoundUserGroupException extends CustomException {
-
-    public NotFoundUserGroupException() {
-        super(ExceptionDetails.NOT_FOUND_USERGROUP);
-    }
-}
